requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Speech to Text
openai-whisper==20231117
SpeechRecognition==3.10.0
# Alternative: google-cloud-speech==2.24.0

# Text to Speech
pyttsx3==2.90
gTTS==2.4.0
elevenlabs==0.2.27
# Alternative: google-cloud-texttospeech==2.16.0

# LLM Providers
google-generativeai==0.3.2  # Gemini
openai==1.6.1  # Optional
anthropic==0.8.1  # Optional

# Audio handling
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
pydub==0.25.1

# Utilities
aiohttp==3.9.1
asyncio==3.4.3
loguru==0.7.2
tenacity==8.2.3  # For retry logic

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0